+++ components/badge/w-badge.vue
@@ -16,10 +16,6 @@ const props = defineProps({
     type: String,
     default: 'neutral',
     validator: (value) => {
-      if (value === 'notification') {
-        console.warn('Warp Badge: The "notification" variant is deprecated - Do not use!');
-        return true;
-      }
       return ['neutral', 'info', 'positive', 'warning', 'negative', 'disabled', 'price'].includes(value);
     },
   },
@@ -32,13 +28,11 @@ const props = defineProps({
 const badgeClasses = computed(() => [
   ccBadge.base,
   ccBadge[props.variant],
-  {
-    [ccBadge.positionBase]: props.position,
-    [ccBadge.positionTL]: props.position === 'top-left',
-    [ccBadge.positionTR]: props.position === 'top-right',
-    [ccBadge.positionBR]: props.position === 'bottom-right',
-    [ccBadge.positionBL]: props.position === 'bottom-left',
-  },
+  !!props.position && ccBadge.positionBase,
+  props.position === 'top-left' && ccBadge.positionTL,
+  props.position === 'top-right' && ccBadge.positionTR,
+  props.position === 'bottom-right' && ccBadge.positionBR,
+  props.position === 'bottom-left' && ccBadge.positionBL,
 ]);
 </script>