+++ components/button/w-button.vue
@@ -6,11 +6,12 @@ import { button as ccButton } from '@warp-ds/css/component-classes';
 
 import { activateI18n } from '../util/i18n';
 
+import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
 
-activateI18n(enMessages, nbMessages, fiMessages);
+activateI18n(enMessages, nbMessages, fiMessages, daMessages);
 
 const ariaValueText = i18n._({
   id: 'button.aria.loading',
@@ -38,53 +39,59 @@ const props = defineProps({
 
 const defaultVariant = props.secondary || !buttonVariants.find((b) => !!props[b]);
 
-const buttonClass = computed(() => ({
-  [ccButton.secondary]: defaultVariant && !props.small && !props.quiet && !props.loading,
-  [ccButton.secondarySmall]: defaultVariant && props.small && !props.quiet && !props.loading,
-  [ccButton.secondarySmallLoading]: defaultVariant && props.small && !props.quiet && props.loading,
-  [ccButton.secondarySmallQuiet]: defaultVariant && props.small && props.quiet && !props.loading,
-  [ccButton.secondarySmallQuietLoading]: defaultVariant && props.small && props.quiet && props.loading,
-  [ccButton.secondaryQuiet]: defaultVariant && !props.small && props.quiet && !props.loading,
-  [ccButton.secondaryQuietLoading]: defaultVariant && !props.small && props.quiet && props.loading,
-  [ccButton.secondaryLoading]: defaultVariant && !props.small && !props.quiet && props.loading,
+const primaryClasses = computed(() => [
+  !props.small && !props.quiet && !props.loading && ccButton.primary,
+  props.small && !props.quiet && !props.loading && ccButton.primarySmall,
+  props.small && props.quiet && !props.loading && ccButton.primarySmallQuiet,
+  props.small && props.loading && (props.quiet ? ccButton.primarySmallQuietLoading : ccButton.primarySmallLoading),
+  !props.small && props.quiet && !props.loading && ccButton.primaryQuiet,
+  !props.small && props.loading && (props.quiet ? ccButton.primaryQuietLoading : ccButton.primaryLoading),
+]);
 
-  [ccButton.primary]: props.primary && !props.small && !props.quiet && !props.loading,
-  [ccButton.primarySmall]: props.primary && props.small && !props.quiet && !props.loading,
-  [ccButton.primarySmallQuiet]: props.primary && props.small && props.quiet && !props.loading,
-  [ccButton.primarySmallLoading]: props.primary && props.small && !props.quiet && props.loading,
-  [ccButton.primarySmallQuietLoading]: props.primary && props.small && props.quiet && props.loading,
-  [ccButton.primaryQuiet]: props.primary && !props.small && props.quiet && !props.loading,
-  [ccButton.primaryQuietLoading]: props.primary && !props.small && props.quiet && props.loading,
-  [ccButton.primaryLoading]: props.primary && !props.small && !props.quiet && props.loading,
+const secondaryClasses = computed(() => [
+  !props.small && !props.quiet && !props.loading && ccButton.secondary,
+  props.small && !props.quiet && !props.loading && ccButton.secondarySmall,
+  props.small && props.quiet && !props.loading && ccButton.secondarySmallQuiet,
+  props.small && props.loading && (props.quiet ? ccButton.secondarySmallQuietLoading : ccButton.secondarySmallLoading),
+  !props.small && props.quiet && !props.loading && ccButton.secondaryQuiet,
+  !props.small && props.loading && (props.quiet ? ccButton.secondaryQuietLoading : ccButton.secondaryLoading),
+]);
 
-  [ccButton.utility]: props.utility && !props.small && !props.quiet && !props.loading,
-  [ccButton.utilitySmall]: props.utility && props.small && !props.quiet && !props.loading,
-  [ccButton.utilitySmallQuiet]: props.utility && props.small && props.quiet && !props.loading,
-  [ccButton.utilitySmallLoading]: props.utility && props.small && !props.quiet && props.loading,
-  [ccButton.utilitySmallQuietLoading]: props.utility && props.small && props.quiet && props.loading,
-  [ccButton.utilityQuiet]: props.utility && !props.small && props.quiet && !props.loading,
-  [ccButton.utilityQuietLoading]: props.utility && !props.small && props.quiet && props.loading,
-  [ccButton.utilityLoading]: props.utility && !props.small && !props.quiet && props.loading,
+const utilityClasses = computed(() => [
+  !props.small && !props.quiet && !props.loading && ccButton.utility,
+  props.small && !props.quiet && !props.loading && ccButton.utilitySmall,
+  props.small && props.quiet && !props.loading && ccButton.utilitySmallQuiet,
+  props.small && props.loading && (props.quiet ? ccButton.utilitySmallQuietLoading : ccButton.utilitySmallLoading),
+  !props.small && props.quiet && !props.loading && ccButton.utilityQuiet,
+  !props.small && props.loading && (props.quiet ? ccButton.utilityQuietLoading : ccButton.utilityLoading),
+]);
 
-  [ccButton.negative]: props.negative && !props.small && !props.quiet && !props.loading,
-  [ccButton.negativeSmall]: props.negative && props.small && !props.quiet && !props.loading,
-  [ccButton.negativeSmallQuiet]: props.negative && props.small && props.quiet && !props.loading,
-  [ccButton.negativeSmallLoading]: props.negative && props.small && !props.quiet && props.loading,
-  [ccButton.negativeSmallQuietLoading]: props.negative && props.small && props.quiet && props.loading,
-  [ccButton.negativeQuiet]: props.negative && !props.small && props.quiet && !props.loading,
-  [ccButton.negativeQuietLoading]: props.negative && !props.small && props.quiet && props.loading,
-  [ccButton.negativeLoading]: props.negative && !props.small && !props.quiet && props.loading,
+const negativeClasses = computed(() => [
+  !props.small && !props.quiet && !props.loading && ccButton.negative,
+  props.small && !props.quiet && !props.loading && ccButton.negativeSmall,
+  props.small && props.quiet && !props.loading && ccButton.negativeSmallQuiet,
+  props.small && props.loading && (props.quiet ? ccButton.negativeSmallQuietLoading : ccButton.negativeSmallLoading),
+  !props.small && props.quiet && !props.loading && ccButton.negativeQuiet,
+  !props.small && props.loading && (props.quiet ? ccButton.negativeQuietLoading : ccButton.negativeLoading),
+]);
 
-  [ccButton.pill]: props.pill && !props.small && !props.loading,
-  [ccButton.pillSmall]: props.pill && props.small && !props.loading,
-  [ccButton.pillLoading]: props.pill && !props.small && props.loading,
-  [ccButton.pillSmallLoading]: props.pill && props.small && props.loading,
-  [ccButton.link]: props.link && !props.small,
-  [ccButton.linkSmall]: props.link && props.small,
-  [ccButton.linkAsButton]: !!props.href,
-  [ccButton.fullWidth]: props.fullWidth,
-  [ccButton.contentWidth]: !props.fullWidth,
-}));
+const pillClasses = computed(() => [
+  !props.loading && (!props.small ? ccButton.pill : ccButton.pillSmall),
+  props.loading && (!props.small ? ccButton.pillLoading : ccButton.pillSmallLoading),
+]);
+
+const linkClasses = computed(() => [props.small ? ccButton.linkSmall : ccButton.link]);
+
+const classes = computed(() => [
+  props.primary && primaryClasses.value,
+  defaultVariant && secondaryClasses.value,
+  props.utility && utilityClasses.value,
+  props.negative && negativeClasses.value,
+  props.pill && pillClasses.value,
+  props.link && linkClasses.value,
+  props.href && ccButton.linkAsButton,
+  props.fullWidth ? ccButton.fullWidth : ccButton.contentWidth,
+]);
 
 const saneDefaults = computed(() => ({
   type: props.href ? undefined : attrs.type || 'button',
@@ -97,7 +104,7 @@ export default { name: 'wButton' };
 </script>
 
 <template>
-  <component :is="href ? 'a' : 'button'" :href="href" :class="buttonClass" v-bind="saneDefaults">
+  <component :is="href ? 'a' : 'button'" :href="href" :class="classes" v-bind="saneDefaults">
     <slot>{{ label }}</slot>
     <span v-if="loading" role="progressbar" aria-valuenow="0" :aria-valuetext="ariaValueText" :class="ccButton.a11y" />
   </component>