+++ components/forms/w-textarea.vue
@@ -7,6 +7,16 @@ import { default as wField, fieldProps } from './w-field.vue';
 const p = defineProps(fieldProps);
 const emit = defineEmits(['update:modelValue']);
 const model = createModel({ props: p, emit });
+
+const textAreaClasses = (hasValidationErrors) => [
+  ccInput.base,
+  ccInput.textArea,
+  !!p.placeholder && ccInput.placeholder,
+  !hasValidationErrors && !p.disabled && !p.readOnly && ccInput.default,
+  hasValidationErrors && !p.disabled && !p.readOnly && ccInput.invalid,
+  !hasValidationErrors && p.disabled && !p.readOnly && ccInput.disabled,
+  !hasValidationErrors && !p.disabled && p.readOnly && ccInput.readOnly,
+];
 </script>
 
 <script>
@@ -17,14 +27,7 @@ export default { name: 'wTextarea', inheritAttrs: false };
   <w-field v-bind="{ ...$attrs, ...$props }" #default="{ triggerValidation, aria, hasValidationErrors }">
     <div :class="[ccInput.wrapper]">
       <textarea
-        :class="{
-          [`${ccInput.base} ${ccInput.textArea}`]: true,
-          [ccInput.placeholder]: !!p.placeholder,
-          [ccInput.default]: !hasValidationErrors && !p.disabled && !p.readOnly,
-          [ccInput.invalid]: hasValidationErrors && !p.disabled && !p.readOnly,
-          [ccInput.disabled]: !hasValidationErrors && p.disabled && !p.readOnly,
-          [ccInput.readOnly]: !hasValidationErrors && !p.disabled && p.readOnly,
-        }"
+        :class="textAreaClasses(hasValidationErrors)"
         :disabled="disabled"
         :readOnly="readOnly"
         v-bind="{ ...aria, ...$attrs, class: '' }"