+++ components/modal/w-modal.vue
@@ -10,13 +10,14 @@ import { setup as setupScrollLock, teardown as teardownScrollLock } from 'scroll
 
 import { activateI18n } from '../util/i18n';
 
+import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
 
 import { id } from '#util';
 
-activateI18n(enMessages, nbMessages, fiMessages);
+activateI18n(enMessages, nbMessages, fiMessages, daMessages);
 
 const ariaBack = i18n._({
   id: 'modal.aria.back',
@@ -154,7 +155,7 @@ export default {
         <div
           v-if="showContent"
           ref="modalEl"
-          :class="ccModal.modal"
+          :class="ccModal.base"
           tabindex="-1"
           aria-modal="true"
           aria-labelledby="w-modal-title"